requirements.txt

Flask==2.3.3
Flask-SocketIO==5.3.6
paho-mqtt==1.6.1
mariadb==1.1.8
requests==2.31.0
python-socketio==5.8.0
python-engineio==4.7.1
python-dotenv==1.0.0
pyserial